+++ src/ethereum/info/bobevmInfo.ts
@@ -1,0 +1,111 @@
+import { EdgeCurrencyInfo, EdgeTokenMap } from 'edge-core-js/types'
+
+import { makeOuterPlugin } from '../../common/innerPlugin'
+import { makeMetaTokens } from '../../common/tokenHelpers'
+import type { EthereumTools } from '../EthereumTools'
+import {
+  asEthereumInfoPayload,
+  EthereumFees,
+  EthereumNetworkInfo
+} from '../ethereumTypes'
+import {
+  evmCustomFeeTemplate,
+  evmCustomTokenTemplate,
+  evmMemoOptions,
+  makeEvmDefaultSettings
+} from './ethereumCommonInfo'
+
+const builtinTokens: EdgeTokenMap = {}
+
+// Fees are in Wei
+const defaultNetworkFees: EthereumFees = {
+  default: {
+    baseFee: undefined,
+    baseFeeMultiplier: {
+      lowFee: '1',
+      standardFeeLow: '1.25',
+      standardFeeHigh: '1.5',
+      highFee: '1.75'
+    },
+    gasLimit: {
+      regularTransaction: '21000',
+      tokenTransaction: '300000',
+      minGasLimit: '21000'
+    },
+    gasPrice: {
+      lowFee: '1000000001',
+      standardFeeLow: '40000000001',
+      standardFeeHigh: '300000000001',
+      standardFeeLowAmount: '100000000000000000',
+      standardFeeHighAmount: '10000000000000000000',
+      highFee: '40000000001',
+      minGasPrice: '100000'
+    },
+    minPriorityFee: '2000000000'
+  }
+}
+
+const networkInfo: EthereumNetworkInfo = {
+  networkAdapterConfigs: [
+    {
+      type: 'rpc',
+      servers: ['https://rpc.gobob.xyz']
+    },
+    {
+      type: 'evmscan',
+      servers: ['https://explorer.gobob.xyz']
+    }
+  ],
+  uriNetworks: [],
+  ercTokenStandard: 'ERC20',
+  chainParams: {
+    chainId: 60808,
+    name: 'BOB'
+  },
+  optimismRollup: true,
+  supportsEIP1559: true,
+  hdPathCoinType: 60,
+  amberDataBlockchainId: '',
+  pluginMnemonicKeyName: 'bobevmMnemonic',
+  pluginRegularKeyName: 'bobevmKey',
+  ethGasStationUrl: null,
+  defaultNetworkFees
+}
+
+const currencyInfo: EdgeCurrencyInfo = {
+  canReplaceByFee: true,
+  currencyCode: 'ETH',
+  customFeeTemplate: evmCustomFeeTemplate,
+  customTokenTemplate: evmCustomTokenTemplate,
+  displayName: 'BOB',
+  memoOptions: evmMemoOptions,
+  pluginId: 'bobevm',
+  walletType: 'wallet:bobevm',
+
+  // Explorers:
+  addressExplorer: 'https://explorer.gobob.xyz/address/%s',
+  transactionExplorer: 'https://explorer.gobob.xyz/tx/%s',
+
+  denominations: [
+    {
+      name: 'ETH',
+      multiplier: '1000000000000000000',
+      symbol: 'Ξ'
+    }
+  ],
+
+  // Deprecated:
+  defaultSettings: makeEvmDefaultSettings(networkInfo),
+  metaTokens: makeMetaTokens(builtinTokens)
+}
+
+export const bobevm = makeOuterPlugin<EthereumNetworkInfo, EthereumTools>({
+  builtinTokens,
+  currencyInfo,
+  infoPayloadCleaner: asEthereumInfoPayload,
+  networkInfo,
+
+  async getInnerPlugin() {
+    return await import('../EthereumTools')
+  }
+})