+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
+- added: Build On Bitcoin (BOB) EVM chain
 - added: (Zcash/Piratechain) Cache addresses on local data
+- added: (Ethereum Classic) Add evmscan provider for transaction history
 
 ## 4.6.0 (2024-05-30)